openai_requirement_modeling/requirements.txt
openai>=1.0.0
pyyaml>=6.0
requests>=2.28.0
typing-extensions>=4.0.0 